While loops and if/else statements
They're implemented as I'd expect them to be, pretty standard stuff. I can't say I like the syntax, but I assume it gets useful down the line as things get more complicated. 

Used an if/else statement to check if variable "name" == "Stephen" or not. 
Used a while loop to count to 10. 

Gonna add comments in the next update.